+++ tests/subscriber.cpp
@@ -16,8 +16,6 @@ BOOST_AUTO_TEST_CASE(construction)
     BOOST_CHECK_NO_THROW(
         zeroeq::Subscriber subscriber(test::buildUniqueSession()));
     BOOST_CHECK_NO_THROW(zeroeq::Subscriber(zeroeq::URI("localhost:1234")));
-    BOOST_CHECK_NO_THROW(
-        zeroeq::Subscriber(zeroeq::URI("localhost"), zeroeq::DEFAULT_SESSION));
 
     zeroeq::Subscriber shared;
     BOOST_CHECK_NO_THROW(zeroeq::Subscriber((zeroeq::Receiver&)shared));
@@ -25,10 +23,6 @@ BOOST_AUTO_TEST_CASE(construction)
         zeroeq::Subscriber(test::buildUniqueSession(), shared));
     BOOST_CHECK_NO_THROW(
         zeroeq::Subscriber(zeroeq::URI("localhost:1234"), shared));
-    BOOST_CHECK_NO_THROW(zeroeq::Subscriber(zeroeq::URI("localhost"),
-                                            zeroeq::DEFAULT_SESSION, shared));
-    BOOST_CHECK_NO_THROW(zeroeq::Subscriber(zeroeq::URI("localhost:1234"),
-                                            zeroeq::DEFAULT_SESSION, shared));
 }
 
 BOOST_AUTO_TEST_CASE(invalid_construction)
@@ -38,11 +32,6 @@ BOOST_AUTO_TEST_CASE(invalid_construction)
     BOOST_CHECK_THROW(zeroeq::Subscriber(""), std::runtime_error);
     BOOST_CHECK_THROW(zeroeq::Subscriber(zeroeq::URI("localhost")),
                       std::runtime_error);
-    BOOST_CHECK_THROW(zeroeq::Subscriber(zeroeq::URI("localhost"),
-                                         zeroeq::NULL_SESSION),
-                      std::runtime_error);
-    BOOST_CHECK_THROW(zeroeq::Subscriber(zeroeq::URI("localhost"), ""),
-                      std::runtime_error);
 
     zeroeq::Subscriber shared;
     BOOST_CHECK_THROW(zeroeq::Subscriber subscriber(zeroeq::NULL_SESSION,
@@ -50,11 +39,6 @@ BOOST_AUTO_TEST_CASE(invalid_construction)
                       std::runtime_error);
     BOOST_CHECK_THROW(zeroeq::Subscriber("", shared), std::runtime_error);
     BOOST_CHECK_THROW(zeroeq::Subscriber(zeroeq::URI("localhost"), shared),
-                      std::runtime_error);
-    BOOST_CHECK_THROW(zeroeq::Subscriber(zeroeq::URI("localhost"),
-                                         zeroeq::NULL_SESSION, shared),
-                      std::runtime_error);
-    BOOST_CHECK_THROW(zeroeq::Subscriber(zeroeq::URI("localhost"), "", shared),
                       std::runtime_error);
 }